+++ test/index.test.js
@@ -2,7 +2,7 @@
 // test modules
 const single = require("./single");
 const multiple = require("./multiple");
-const { desktop, mobile } = require("./pageSpeed");
+const pageSpeed = require("./pageSpeed");
 const githubInfo = require("./githubinfo");
 const animeSearch = require("./githubinfo");
 const {
@@ -15,13 +15,13 @@ const {
   biosInfo
 } = require("./hardware");
 
+const cryptoMarket = require("./crypto");
+
 test('single stack', async () => await single("http://example.com"), 34000);
 
 test('multiple stack', async () => await multiple(["http://example.com", "https://www.wappalyzer.com/"]), 40000);
 
-test('desktop speed', async () => await desktop("http://example.com"), 30000);
-
-test('mobile speed', async () => await mobile("http://example.com"), 30000);
+test('desktop speed', async () => await pageSpeed("http://example.com"), 30000);
 
 test('github info', async () => await githubInfo("faztweb"));
 
@@ -40,3 +40,5 @@ test('displayInfo', async () => await displayInfo());
 test('controllerInfo', async () => await controllerInfo());
 
 test('biosInfo', async () => await biosInfo());
+
+test('crypto market', async () => await cryptoMarket());