+++ test/pageSpeed.js
@@ -1,38 +1,33 @@
-const axios = require("axios").default
+const axios = require("axios").default;
 
-const mobile = async (url) => {
-  let testReturn;
-  const res = await axios.get(
-    `https://www.googleapis.com/pagespeedonline/v5/runPagespeed?url=${url}&key=AIzaSyBEDaW4FxSZ2s1vz5CdD5Ai6PGZGdAzij0&strategy=mobile`
-  )
+const pageSpeed = async (url) => {
+  const testReturn = [];
+  const resMobile = await axios.get("https://www.googleapis.com/pagespeedonline/v5/runPagespeed", {
+    params: {
+      url,
+      key: "AIzaSyBEDaW4FxSZ2s1vz5CdD5Ai6PGZGdAzij0",
+      strategy: "mobile"
+    }
+  });
+
+  const resDesktop = await axios.get("https://www.googleapis.com/pagespeedonline/v5/runPagespeed", {
+    params: {
+      url,
+      key: "AIzaSyBEDaW4FxSZ2s1vz5CdD5Ai6PGZGdAzij0",
+      strategy: "desktop"
+    }
+  });
   
   try {
-    const movil = res.data.lighthouseResult.categories.performance.score * 100
+    const movil = resMobile.data.lighthouseResult.categories.performance.score * 100;
+    const desktop = resDesktop.data.lighthouseResult.categories.performance.score * 100;
 
-    testReturn = movil;
+    testReturn.push(movil);
+    testReturn.push(desktop);
   } catch (err) {
-    testReturn = err;
+    testReturn.push(err.message);
   }
   return testReturn;
-}
+};
 
-const desktop = async (url) => {
-  let testReturn;
-  const res = await axios.get(
-    `https://www.googleapis.com/pagespeedonline/v5/runPagespeed?url=${url}&key=AIzaSyBEDaW4FxSZ2s1vz5CdD5Ai6PGZGdAzij0&strategy=desktop`
-  ) 
-
-  try {
-    const desktop = res.data.lighthouseResult.categories.performance.score * 100
-
-    testReturn = desktop;
-  } catch (err) {
-    testReturn =  err.message;
-  }
-  return testReturn;
-}
-
-module.exports = {
-  mobile,
-  desktop
-}
+module.exports = pageSpeed;